DROID-3347 object view mapping refactoring
